Fix module import error by ensuring isAnagram.js is correctly exported and imported in test.js
