--- 01-js/easy/anagram.js
@@ -1,11 +1,0 @@
-/*
-  Write a function `isAnagram` which takes 2 parameters and returns true/false if those are anagrams or not.
-  What's Anagram?
-  - A word, phrase, or name formed by rearranging the letters of another, such as spar, formed from rasp.
-*/
-
-function isAnagram(str1, str2) {
-  
-}
-
-module.exports = isAnagram;
