+++ 01-js/easy/test.js
@@ -1,0 +1,4 @@
+﻿const isAnagram = require('./isAnagram');
+
+console.log(isAnagram("spar", "rasp"));        // true
+console.log(isAnagram("hello", "world"));      // false
